✨ Add release workflow and goreleaser configuration files
Added a GitHub Actions workflow file 'release.yml' to run Goreleaser
when pushing a tag in the SEMVER format, and a configuration file for
Goreleaser '.goreleaser.yaml' with some sensible defaults.
Also added a new file to .gitignore 'dist/'.

The automation of releases will reduce the workload of manually creating
binary files for several operational systems and architectures, and it
also helps ensure uniformity across builds.